+++ src/coord.h
@@ -25,6 +25,6 @@ void set_coord_y(coord* point, double y);
 float distance(coord* p1, coord* p2);
 float distance_euclid(coord* p1, coord* p2);
 bool excl_carre(coord* point, coord* dep, coord* arr, int marge);
-bool excl_ovale(coord* point, coord* dep, coord* arr, int marge);
+bool excl_ovale(coord* point, coord* dep, coord* arr, double marge);
 
 #endif
